Fix #40.
When processing the queue, only allow one redeployment for each player. We pick one arbitrarily, so we make a best effort to prevent players from getting into this situation by marking them as having completed the phase after issuing an order. We then ignore that player if they attempt to issue a second deployment, with a 409 HTTP error which we can display to the user in a friendly way.